Add files via upload
This repository contains a comprehensive SQL-based exploratory data analysis (EDA) project on the layoffs dataset. The project began with data cleaning, where the dataset was prepared and uploaded. Following this, an in-depth exploration was conducted to derive meaningful insights. All SQL queries utilized throughout the process are documented within the repository.